Bugfixes, Wrappers, CI/CD and Test Automation (#5)
So I reworked the module to enable CI/CD:

Added Invoke-Build, BuildHelpers, PlatyPS, Pester and PSScriptanalyzer.

I wrote Pester tests for all functionality except the two Invoke functions. (which will be in the next iteration).

Because of these changes i had to re-arrange the Module which will BREAK existing installations. I'm sorry for that but this will make sure the future of this repo will be ensured.

Also wrote some wrappers for calling functions and generally improved the coding quality because of Pester.